✨🏷️🚧✅ Object::Unit now use a long and not a int. First implementation of numbers handling like in RFC8259. Still need to deal with fractionnal part.

+++ includes/miscellaneous/JSON/spk_JSON_object.hpp
@@ -13,7 +13,7 @@ namespace spk
 		class Object
 		{
 		public:
-			using Unit = std::variant<bool, int, double, std::wstring, Object *, std::nullptr_t>;
+			using Unit = std::variant<bool, long, double, std::wstring, Object *, std::nullptr_t>;
 			using ContentType = std::variant<Unit, std::map<std::wstring, Object *>, std::vector<Object *>>;
 
 		private:
